document previews batch get

+++ src/main/kotlin/net/kvibews/config/Properties.kt
@@ -31,7 +31,7 @@ class ApplicationProperties(
 ) {
     class Operation(
         var maxNumberOfRetries: Int = 4,
-        var staleDocumentExpiry: Int = 7200
+        var staleDocumentExpiry: Long = 7200
     )
 }
